+++ contracts/finality/src/exec/public_randomness.rs
@@ -3,6 +3,7 @@ use crate::error::ContractError;
 use crate::msg::BabylonMsg;
 use crate::state::config::get_config;
 use crate::state::public_randomness::{insert_pub_rand_commit, PubRandCommit};
+use crate::state::rate_limiting::check_rate_limit_and_accumulate;
 use crate::utils::get_fp_rand_commit_context_v0;
 use crate::utils::query_finality_provider;
 use babylon_bindings::BabylonQuery;
@@ -58,6 +59,10 @@ pub fn handle_public_randomness_commit(
     ensure_fp_exists_and_not_slashed(deps.as_ref(), fp_btc_pk_hex)?;
 
     let fp_btc_pk = hex::decode(fp_btc_pk_hex)?;
+
+    // Ensure rate limiting and accumulate; this step will fail if the rate limit is exceeded
+    check_rate_limit_and_accumulate(deps.storage, env, &fp_btc_pk)?;
+
     let context = get_fp_rand_commit_context_v0(env)?;
     // Verify signature over the list
     verify_commitment_signature(
@@ -178,9 +183,12 @@ fn ensure_fp_exists_and_not_slashed(
 pub(crate) mod tests {
     use super::*;
     use crate::contract::tests::mock_deps_babylon;
-    use crate::state::config::{Config, CONFIG};
+    use crate::state::config::{Config, RateLimitingConfig, CONFIG};
     use babylon_test_utils::get_public_randomness_commitment;
     use cosmwasm_std::testing::mock_env;
+
+    const MAX_MSGS_PER_INTERVAL: u32 = 100;
+    const RATE_LIMITING_INTERVAL: u64 = 10000;
 
     #[test]
     fn verify_commitment_signature_works() {
@@ -215,6 +223,10 @@ pub(crate) mod tests {
         let config = Config {
             bsn_id: format!("test-{}", get_random_u64()),
             min_pub_rand: get_random_u64(),
+            rate_limiting: RateLimitingConfig {
+                max_msgs_per_interval: MAX_MSGS_PER_INTERVAL,
+                block_interval: RATE_LIMITING_INTERVAL,
+            },
             bsn_activation_height: 1,
             finality_signature_interval: 1,
         };
@@ -250,6 +262,10 @@ pub(crate) mod tests {
         let config = Config {
             bsn_id: format!("test-{}", get_random_u64()),
             min_pub_rand: get_random_u64(),
+            rate_limiting: RateLimitingConfig {
+                max_msgs_per_interval: MAX_MSGS_PER_INTERVAL,
+                block_interval: RATE_LIMITING_INTERVAL,
+            },
             bsn_activation_height: 1,
             finality_signature_interval: 1,
         };
@@ -320,6 +336,10 @@ pub(crate) mod tests {
         let config = Config {
             bsn_id: format!("test-{}", get_random_u64()),
             min_pub_rand: get_random_u64(),
+            rate_limiting: RateLimitingConfig {
+                max_msgs_per_interval: MAX_MSGS_PER_INTERVAL,
+                block_interval: RATE_LIMITING_INTERVAL,
+            },
             bsn_activation_height: 1,
             finality_signature_interval: 1,
         };
@@ -349,6 +369,10 @@ pub(crate) mod tests {
         let config = Config {
             bsn_id: format!("test-{}", get_random_u64()),
             min_pub_rand: get_random_u64(),
+            rate_limiting: RateLimitingConfig {
+                max_msgs_per_interval: MAX_MSGS_PER_INTERVAL,
+                block_interval: RATE_LIMITING_INTERVAL,
+            },
             bsn_activation_height: 1,
             finality_signature_interval: 1,
         };
@@ -417,6 +441,10 @@ pub(crate) mod tests {
         let config = Config {
             bsn_id: format!("test-{}", get_random_u64()),
             min_pub_rand: get_random_u64(),
+            rate_limiting: RateLimitingConfig {
+                max_msgs_per_interval: MAX_MSGS_PER_INTERVAL,
+                block_interval: RATE_LIMITING_INTERVAL,
+            },
             bsn_activation_height: 1,
             finality_signature_interval: 1,
         };
@@ -459,6 +487,10 @@ pub(crate) mod tests {
         let config = Config {
             bsn_id: format!("test-{}", get_random_u64()),
             min_pub_rand,
+            rate_limiting: RateLimitingConfig {
+                max_msgs_per_interval: MAX_MSGS_PER_INTERVAL,
+                block_interval: RATE_LIMITING_INTERVAL,
+            },
             bsn_activation_height: 1,
             finality_signature_interval: 1,
         };